+++ package/src/firebaseClientLike/index.js
@@ -6,8 +6,11 @@
 *
 * Provides support for using Firebase clients in eg. Cloud Functions testing.
 */
-import { projectId } from "../config.js"
 
+/*** disabled
+import { projectId } from "../config.js"
 export {
   projectId   // pass back (to tests) the project ID the test application has provided (in Global Setup)
-}
+}***/
+
+export { httpsCallable } from './httpsCallable.js'