+++ package/src/firestoreREST/commit.js
@@ -50,7 +50,7 @@ async function commit_v1(token, writes) {   // (string|null, Array of Write) => 
 
   //console.log("!!! Fetch", { token, uri, method, body } );  // DEBUG
 
-  const res = await fetch(uri, {method, headers: token ? { ["Authorization"]: `Bearer ${token}` } : {}, body })
+  const res = await fetch(uri, {method, headers: token ? { "Authorization": `Bearer ${token}` } : {}, body })
   const status = res.status;
 
   // Access: